rename pkg dir
