+++ pypeit/pkg/exceptions.py
